Move all management command routines to hs_core/management/utils.py
