+++ hs_core/management/commands/repair_resource.py
@@ -15,7 +15,8 @@ This checks that:
 
 from django.core.management.base import BaseCommand
 from hs_core.models import BaseResource
-from hs_core.management.utils import ingest_irods_files
+from hs_core.hydroshare.utils import get_resource_by_shortkey
+from hs_core.management.utils import repair_resource
 
 import logging
 
@@ -38,139 +39,50 @@ class Command(BaseCommand):
 
     def handle(self, *args, **options):
 
+        logger = logging.getLogger(__name__)
         log_errors = options['log']
         echo_errors = not options['log']
-        logger = logging.getLogger(__name__)
 
         if len(options['resource_ids']) > 0:  # an array of resource short_id to check.
             for rid in options['resource_ids']:
                 try:
-                    r = BaseResource.objects.get(short_id=rid)
-                except BaseResource.DoesNotExist:
-                    msg = "Resource with id {} not found in Django Resources".format(rid)
-                    if log_errors:
-                        logger.error(msg)
-                    if echo_errors:
-                        print(msg)
-                    continue  # next resource
-
-                try:
-                    resource = r.get_content_model()
-                except Exception as e:
-                    msg = "resource {} has no proxy class: {}"\
-                          .format(r.short_id, e.value)
-                    if log_errors:
-                        logger.error(msg)
-                    if echo_errors:
-                        print(msg)
+                    resource = get_resource_by_shortkey(rid)
+                except BaseResource.NotFoundException:
+                    msg = "resource {} not found".format(rid)
+                    print(msg)
+                    continue
+                _, count = repair_resource(resource,
+                                           echo_errors=echo_errors,
+                                           log_errors=log_errors,
+                                           return_errors=False)
+                if count:
                     msg = "... affected resource {} has type {}, title '{}'"\
-                          .format(r.short_id, r.resource_type, r.title)
+                          .format(resource.short_id, resource.resource_type,
+                                  resource.title.encode('string-escape'))
                     if log_errors:
                         logger.info(msg)
                     if echo_errors:
                         print(msg)
-                    continue  # next resource
-
-                print("REPAIRING RESOURCE {}".format(rid))
-
-                # ingest any dangling iRODS files that you can
-                # Do this before check because otherwise, errors get printed twice
-                # TODO: This does not currently work properly for composite resources
-                # if resource.resource_type == 'CompositeResource' or \
-                if resource.resource_type == 'GenericResource' or \
-                   resource.resource_type == 'ModelInstanceResource' or \
-                   resource.resource_type == 'ModelProgramResource':
-                    _, count = ingest_irods_files(resource,
-                                                  logger,
-                                                  stop_on_error=False,
-                                                  echo_errors=True,
-                                                  log_errors=False,
-                                                  return_errors=False)
-                    if count:
-                        print("... affected resource {} has type {}, title '{}'"
-                              .format(resource.short_id, resource.resource_type, resource.title))
-                _, count = resource.check_irods_files(stop_on_error=False,
-                                                      echo_errors=True,
-                                                      log_errors=False,
-                                                      return_errors=False,
-                                                      clean_irods=False,
-                                                      clean_django=True,
-                                                      sync_ispublic=True)
-                if count:
-                    print("... affected resource {} has type {}, title '{}'"
-                          .format(resource.short_id, resource.resource_type, resource.title))
-
-            # NOT NEEDED after composite resource phase II
-            # if resource.resource_type == 'CompositeResource':
-            #     count = 0
-            #     for res_file in resource.files.all():
-            #         if not res_file.has_logical_file:
-            #             count += 1
-            #             print("Logical file missing for file {} (CREATING)"
-            #                   .format(res_file.short_path))
-            #     resource.set_default_logical_file()
-            if count:
-                print("... affected resource {} has type {}, title '{}'"
-                      .format(resource.short_id, resource.resource_type, resource.title))
 
         else:  # check all resources
             print("REPAIRING ALL RESOURCES")
             for r in BaseResource.objects.all():
                 try:
-                    resource = r.get_content_model()
-                except Exception as e:
-                    msg = "resource {} has no proxy class: {}".format(r.short_id, e.value)
-                    if echo_errors:
-                        print(msg)
-                    if log_errors:
-                        logger.error(msg)
+                    resource = get_resource_by_shortkey(r.short_id)
+                except BaseResource.NotFoundException:
+                    msg = "resource {} not found".format(rid)
+                    print(msg)
+                    continue
+
+                _, count = repair_resource(resource,
+                                           echo_errors=echo_errors,
+                                           log_errors=log_errors,
+                                           return_errors=False)
+                if count:
                     msg = "... affected resource {} has type {}, title '{}'"\
-                          .format(r.short_id, r.resource_type, r.title)
+                          .format(resource.short_id, resource.resource_type,
+                                  resource.title.encode('string-escape'))
                     if log_errors:
                         logger.info(msg)
                     if echo_errors:
                         print(msg)
-                    continue  # next resource
-
-                # ingest any dangling iRODS files that you can
-                # Do this before check because otherwise, errors get printed twice
-                # TODO: This does not currently work properly for composite resources
-                # if resource.resource_type == 'CompositeResource' or \
-                if resource.resource_type == 'GenericResource' or \
-                   resource.resource_type == 'ModelInstanceResource' or \
-                   resource.resource_type == 'ModelProgramResource':
-                    _, count = ingest_irods_files(resource,
-                                                  logger,
-                                                  stop_on_error=False,
-                                                  echo_errors=True,
-                                                  log_errors=False,
-                                                  return_errors=False)
-                    if count:
-                        print("... affected resource {} has type {}, title '{}'"
-                              .format(resource.short_id, resource.resource_type, resource.title))
-
-                # clean up Django references to non-existent iRODS files
-                _, count = resource.check_irods_files(stop_on_error=False,
-                                                      echo_errors=True,
-                                                      log_errors=False,
-                                                      return_errors=False,
-                                                      clean_irods=False,
-                                                      clean_django=True,
-                                                      sync_ispublic=True)
-                if count:
-                    print("... affected resource {} has type {}, title '{}'"
-                          .format(resource.short_id, resource.resource_type, resource.title))
-
-                # NOT NEEDED after composite resource phase II
-                # if resource.resource_type == 'CompositeResource':
-                #     count = 0
-                #     for res_file in resource.files.all():
-                #         if not res_file.has_logical_file:
-                #             count += 1
-                #             print("Resource {}: logical file missing for file {} (CREATING)"
-                #                   .format(resource.short_id, res_file.short_path))
-                #     resource.set_default_logical_file()
-
-                if count:
-                    print("... affected resource {} has type {}, title '{}'"
-                          .format(resource.short_id, resource.resource_type, resource.title))
