+++ hs_core/management/utils.py
@@ -11,6 +11,7 @@ If a file in iRODS is not present in Django, it attempts to register that file i
 """
 
 import json
+import os
 from django.conf import settings
 from django.contrib.sites.models import Site
 from django.core.exceptions import ValidationError
@@ -27,21 +28,375 @@ from hs_core.views.utils import link_irods_file_to_django
 import logging
 
 
-def ingest_irods_files(self,
+def check_relations(resource):
+    """Check for dangling relations due to deleted resource files.
+
+    :param resource: resource to check
+    """
+    for r in resource.metadata.relations.all():
+        if r.value.startswith('http://www.hydroshare.org/resource/'):
+            target = r.value[len('http://www.hydroshare.org/resource/'):].rstrip('/')
+            try:
+                get_resource_by_shortkey(target, or_404=False)
+            except BaseResource.DoesNotExist:
+                print("relation {} {} {} (this does not exist)"
+                      .format(resource.short_id, r.type, target))
+
+
+def fix_irods_user_paths(resource, log_actions=True, echo_actions=False, return_actions=False):
+    """Move iRODS user paths to the locations specified in settings.
+
+    :param resource: resource to check
+    :param log_actions: whether to log actions to Django log
+    :param echo_actions: whether to print actions on stdout
+    :param return_actions: whether to collect actions in an array and return them.
+
+    This is a temporary fix to the user resources, which are currently stored like
+    federated resources but whose paths are dynamically determined. This function points
+    the paths for user-level resources to where they are stored in the current environment,
+    as specified in hydroshare/local_settings.py.
+
+    * This only does something if the environment is not a production environment.
+    * It is idempotent, in the sense that it can be repeated more than once without problems.
+    * It must be done once whenever the django database is reloaded.
+    * It does not check whether the paths exist afterward. This is done by check_irods_files.
+    """
+    logger = logging.getLogger(__name__)
+    actions = []
+    ecount = 0
+
+    # location of the user files in production
+    defaultpath = getattr(settings, 'HS_USER_ZONE_PRODUCTION_PATH',
+                          '/hydroshareuserZone/home/localHydroProxy')
+    # where resource should be found; this is equal to the default path in production
+    userpath = '/' + os.path.join(
+        getattr(settings, 'HS_USER_IRODS_ZONE', 'hydroshareuserZone'),
+        'home',
+        getattr(settings, 'HS_LOCAL_PROXY_USER_IN_FED_ZONE', 'localHydroProxy'))
+
+    msg = "fix_irods_user_paths: user path is {}".format(userpath.encode('string-escape'))
+    if echo_actions:
+        print(msg)
+    if log_actions:
+        logger.info(msg)
+    if return_actions:
+        actions.append(msg)
+
+    # only take action if you find a path that is a default user path and not in production
+    if resource.resource_federation_path == defaultpath and userpath != defaultpath:
+        msg = "fix_irods_user_paths: mapping existing user federation path {} to {}"\
+              .format(resource.resource_federation_path, userpath.encode('string-escape'))
+        if echo_actions:
+            print(msg)
+        if log_actions:
+            logger.info(msg)
+        if return_actions:
+            actions.append(msg)
+
+        resource.resource_federation_path = userpath
+        resource.save()
+        for f in resource.files.all():
+            path = f.storage_path
+            if path.startswith(defaultpath):
+                newpath = userpath + path[len(defaultpath):]
+                f.set_storage_path(newpath, test_exists=False)  # does implicit save
+                ecount += 1
+                msg = "fix_irods_user_paths: rewriting {} to {}"\
+                    .format(path.encode('string-escape'),
+                            newpath.encode('string-escape'))
+                if echo_actions:
+                    print(msg)
+                if log_actions:
+                    logger.info(msg)
+                if return_actions:
+                    actions.append(msg)
+            else:
+                msg = ("fix_irods_user_paths: ERROR: malformed path {} in resource" +
+                       " {} should start with {}; cannot convert")\
+                    .format(path.encode('string-escape'), resource.short_id, defaultpath)
+                if echo_actions:
+                    print(msg)
+                if log_actions:
+                    logger.error(msg)
+                if return_actions:
+                    actions.append(msg)
+
+    if ecount > 0:  # print information about the affected resource (not really an error)
+        msg = "fix_irods_user_paths: affected resource {} type is {}, title is '{}'"\
+            .format(resource.short_id, resource.resource_type,
+                    resource.metadata.title.value.encode('string-escape'))
+        if log_actions:
+            logger.info(msg)
+        if echo_actions:
+            print(msg)
+        if return_actions:
+            actions.append(msg)
+
+    return actions, ecount  # empty unless return_actions=True
+
+
+def check_irods_files(resource, stop_on_error=False, log_errors=True,
+                      echo_errors=False, return_errors=False,
+                      sync_ispublic=False, clean_irods=False, clean_django=False):
+    """Check whether files in resource.files and on iRODS agree.
+
+    :param resource: resource to check
+    :param stop_on_error: whether to raise a ValidationError exception on first error
+    :param log_errors: whether to log errors to Django log
+    :param echo_errors: whether to print errors on stdout
+    :param return_errors: whether to collect errors in an array and return them.
+    :param sync_ispublic: whether to repair deviations between ResourceAccess.public
+           and AVU isPublic
+    :param clean_irods: whether to delete files in iRODs that are not in Django
+    :param clean_django: whether to delete files in Django that are not in iRODs
+    """
+    from hs_core.hydroshare.resource import delete_resource_file
+
+    logger = logging.getLogger(__name__)
+    istorage = resource.get_irods_storage()
+    errors = []
+    ecount = 0
+    defaultpath = getattr(settings, 'HS_USER_ZONE_PRODUCTION_PATH',
+                          '/hydroshareuserZone/home/localHydroProxy')
+
+    # skip federated resources if not configured to handle these
+    if resource.is_federated and not settings.REMOTE_USE_IRODS:
+        msg = "check_irods_files: skipping check of federated resource {} in unfederated mode"\
+            .format(resource.short_id)
+        if echo_errors:
+            print(msg)
+        if log_errors:
+            logger.info(msg)
+
+    # skip resources that do not exist in iRODS
+    elif not istorage.exists(resource.root_path):
+            msg = "root path {} does not exist in iRODS".format(resource.root_path)
+            ecount += 1
+            if echo_errors:
+                print(msg)
+            if log_errors:
+                logger.error(msg)
+            if return_errors:
+                errors.append(msg)
+
+    else:
+        # Step 1: repair irods user file paths if necessary
+        if clean_irods or clean_django:
+            # fix user paths before check (required). This is an idempotent step.
+            if resource.resource_federation_path == defaultpath:
+                error2, ecount2 = fix_irods_user_paths(resource,
+                                                       log_actions=log_errors,
+                                                       echo_actions=echo_errors,
+                                                       return_actions=False)
+                errors.extend(error2)
+                ecount += ecount2
+
+        # Step 2: does every file in Django refer to an existing file in iRODS?
+        for f in resource.files.all():
+            if not istorage.exists(f.storage_path):
+                ecount += 1
+                msg = "check_irods_files: django file {} does not exist in iRODS"\
+                    .format(f.storage_path)
+                if clean_django:
+                    delete_resource_file(resource.short_id, f.short_path, resource.creator,
+                                         delete_logical_file=False)
+                    msg += " (DELETED FROM DJANGO)"
+                if echo_errors:
+                    print(msg)
+                if log_errors:
+                    logger.error(msg)
+                if return_errors:
+                    errors.append(msg)
+                if stop_on_error:
+                    raise ValidationError(msg)
+
+        # Step 3: for composite resources, does every composite metadata file exist?
+        from hs_composite_resource.models import CompositeResource as CR
+        if isinstance(resource, CR):
+            for lf in resource.logical_files:
+                if not istorage.exists(lf.metadata_file_path):
+                    ecount += 1
+                    msg = "check_irods_files: logical metadata file {} does not exist in iRODS"\
+                        .format(lf.metadata_file_path)
+                    if echo_errors:
+                        print(msg)
+                    if log_errors:
+                        logger.error(msg)
+                    if return_errors:
+                        errors.append(msg)
+                    if stop_on_error:
+                        raise ValidationError(msg)
+                if not istorage.exists(lf.map_file_path):
+                    ecount += 1
+                    msg = "check_irods_files: logical map file {} does not exist in iRODS"\
+                        .format(lf.map_file_path)
+                    if echo_errors:
+                        print(msg)
+                    if log_errors:
+                        logger.error(msg)
+                    if return_errors:
+                        errors.append(msg)
+                    if stop_on_error:
+                        raise ValidationError(msg)
+
+        # Step 4: does every iRODS file correspond to a record in files?
+        error2, ecount2 = __check_irods_directory(resource, resource.file_path, logger,
+                                                  stop_on_error=stop_on_error,
+                                                  log_errors=log_errors,
+                                                  echo_errors=echo_errors,
+                                                  return_errors=return_errors,
+                                                  clean=clean_irods)
+        errors.extend(error2)
+        ecount += ecount2
+
+        # Step 5: check whether the iRODS public flag agrees with Django
+        django_public = resource.raccess.public
+        irods_public = None
+        try:
+            irods_public = resource.getAVU('isPublic')
+        except SessionException as ex:
+            msg = "cannot read isPublic attribute of {}: {}"\
+                .format(resource.short_id, ex.stderr)
+            ecount += 1
+            if log_errors:
+                logger.error(msg)
+            if echo_errors:
+                print(msg)
+            if return_errors:
+                errors.append(msg)
+            if stop_on_error:
+                raise ValidationError(msg)
+
+        if irods_public is not None:
+            # convert to boolean
+            irods_public = str(irods_public).lower() == 'true'
+
+        if irods_public is None or irods_public != django_public:
+            ecount += 1
+            if not django_public:  # and irods_public
+                msg = "check_irods_files: resource {} public in irods, private in Django"\
+                    .format(resource.short_id)
+                if sync_ispublic:
+                    try:
+                        resource.setAVU('isPublic', 'false')
+                        msg += " (REPAIRED IN IRODS)"
+                    except SessionException as ex:
+                        msg += ": (CANNOT REPAIR: {})"\
+                            .format(ex.stderr)
+
+            else:  # django_public and not irods_public
+                msg = "check_irods_files: resource {} private in irods, public in Django"\
+                    .format(resource.short_id)
+                if sync_ispublic:
+                    try:
+                        resource.setAVU('isPublic', 'true')
+                        msg += " (REPAIRED IN IRODS)"
+                    except SessionException as ex:
+                        msg += ": (CANNOT REPAIR: {})"\
+                            .format(ex.stderr)
+
+            if msg != '':
+                if echo_errors:
+                    print(msg)
+                if log_errors:
+                    logger.error(msg)
+                if return_errors:
+                    errors.append(msg)
+                if stop_on_error:
+                    raise ValidationError(msg)
+
+    if ecount > 0:  # print information about the affected resource (not really an error)
+        msg = "check_irods_files: affected resource {} type is {}, title is '{}'"\
+            .format(resource.short_id, resource.resource_type,
+                    resource.metadata.title.value.encode('string-escape'))
+        if log_errors:
+            logger.error(msg)
+        if echo_errors:
+            print(msg)
+        if return_errors:
+            errors.append(msg)
+
+    return errors, ecount  # empty unless return_errors=True
+
+
+def __check_irods_directory(resource, dir, logger,
+                            stop_on_error=False, log_errors=True,
+                            echo_errors=False, return_errors=False,
+                            clean=False):
+    """List a directory and check files there for conformance with django ResourceFiles.
+
+    :param stop_on_error: whether to raise a ValidationError exception on first error
+    :param log_errors: whether to log errors to Django log
+    :param echo_errors: whether to print errors on stdout
+    :param return_errors: whether to collect errors in an array and return them.
+
+    """
+    errors = []
+    ecount = 0
+    istorage = resource.get_irods_storage()
+    try:
+        listing = istorage.listdir(dir)
+        for fname in listing[1]:  # files
+            # do not use os.path.join because fname might contain unicode characters
+            fullpath = dir + '/' + fname
+            found = False
+            for f in resource.files.all():
+                if f.storage_path == fullpath:
+                    found = True
+                    break
+            if not found and not resource.is_aggregation_xml_file(fullpath):
+                ecount += 1
+                msg = "check_irods_files: file {} in iRODs does not exist in Django"\
+                    .format(fullpath.encode('string-escape'))
+                if clean:
+                    try:
+                        istorage.delete(fullpath)
+                        msg += " (DELETED FROM IRODS)"
+                    except SessionException as ex:
+                        msg += ": (CANNOT DELETE: {})"\
+                            .format(ex.stderr)
+                if echo_errors:
+                    print(msg)
+                if log_errors:
+                    logger.error(msg)
+                if return_errors:
+                    errors.append(msg)
+                if stop_on_error:
+                    raise ValidationError(msg)
+
+        for dname in listing[0]:  # directories
+            # do not use os.path.join because paths might contain unicode characters!
+            error2, ecount2 = __check_irods_directory(resource, dir + '/' + dname, logger,
+                                                      stop_on_error=stop_on_error,
+                                                      echo_errors=echo_errors,
+                                                      log_errors=log_errors,
+                                                      return_errors=return_errors,
+                                                      clean=clean)
+            errors.extend(error2)
+            ecount += ecount2
+
+    except SessionException:
+        pass  # not an error not to have a file directory.
+        # Non-existence of files is checked elsewhere.
+
+    return errors, ecount  # empty unless return_errors=True
+
+
+def ingest_irods_files(resource,
                        logger,
                        stop_on_error=False,
                        echo_errors=True,
                        log_errors=False,
                        return_errors=False):
 
-    istorage = self.get_irods_storage()
+    istorage = resource.get_irods_storage()
     errors = []
     ecount = 0
 
     # skip federated resources if not configured to handle these
-    if self.is_federated and not settings.REMOTE_USE_IRODS:
+    if resource.is_federated and not settings.REMOTE_USE_IRODS:
         msg = "ingest_irods_files: skipping ingest of federated resource {} in unfederated mode"\
-            .format(self.short_id)
+            .format(resource.short_id)
         if echo_errors:
             print(msg)
         if log_errors:
@@ -49,8 +404,8 @@ def ingest_irods_files(self,
 
     else:
         # flag non-existent resources in iRODS
-        if not istorage.exists(self.root_path):
-            msg = "root path {} does not exist in iRODS".format(self.root_path)
+        if not istorage.exists(resource.root_path):
+            msg = "root path {} does not exist in iRODS".format(resource.root_path)
             ecount += 1
             if echo_errors:
                 print(msg)
@@ -60,8 +415,8 @@ def ingest_irods_files(self,
                 errors.append(msg)
 
         # flag non-existent file paths in iRODS
-        elif not istorage.exists(self.file_path):
-            msg = "file path {} does not exist in iRODS".format(self.file_path)
+        elif not istorage.exists(resource.file_path):
+            msg = "file path {} does not exist in iRODS".format(resource.file_path)
             ecount += 1
             if echo_errors:
                 print(msg)
@@ -71,8 +426,8 @@ def ingest_irods_files(self,
                 errors.append(msg)
 
         else:
-            return __ingest_irods_directory(self,
-                                            self.file_path,
+            return __ingest_irods_directory(resource,
+                                            resource.file_path,
                                             logger,
                                             stop_on_error=False,
                                             echo_errors=True,
@@ -82,7 +437,7 @@ def ingest_irods_files(self,
     return errors, ecount
 
 
-def __ingest_irods_directory(self,
+def __ingest_irods_directory(resource,
                              dir,
                              logger,
                              stop_on_error=False,
@@ -100,17 +455,18 @@ def __ingest_irods_directory(self,
     """
     errors = []
     ecount = 0
-    istorage = self.get_irods_storage()
+    istorage = resource.get_irods_storage()
     try:
         listing = istorage.listdir(dir)
         for fname in listing[1]:  # files
+            # do not use os.path.join because fname might contain unicode characters
             fullpath = dir + '/' + fname
             found = False
-            for res_file in self.files.all():
+            for res_file in resource.files.all():
                 if res_file.storage_path == fullpath:
                     found = True
 
-            if not found and not self.is_aggregation_xml_file(fullpath):
+            if not found and not resource.is_aggregation_xml_file(fullpath):
                 ecount += 1
                 msg = "ingest_irods_files: file {} in iRODs does not exist in Django (INGESTING)"\
                     .format(fullpath.encode('string-escape'))
@@ -123,11 +479,11 @@ def __ingest_irods_directory(self,
                 if stop_on_error:
                     raise ValidationError(msg)
                 # TODO: does not ingest logical file structure for composite resources
-                link_irods_file_to_django(self, fullpath)
+                link_irods_file_to_django(resource, fullpath)
 
                 # Create required logical files as necessary
-                if self.resource_type == "CompositeResource":
-                    file_type = get_logical_file_type(res=self, user=None,
+                if resource.resource_type == "CompositeResource":
+                    file_type = get_logical_file_type(res=resource, user=None,
                                                       file_id=res_file.pk, fail_feedback=False)
                     if not res_file.has_logical_file and file_type is not None:
                         msg = "ingest_irods_files: setting required logical file for {}"\
@@ -140,7 +496,7 @@ def __ingest_irods_directory(self,
                             errors.append(msg)
                         if stop_on_error:
                             raise ValidationError(msg)
-                        set_logical_file_type(res=self, user=None, file_id=res_file.pk,
+                        set_logical_file_type(res=resource, user=None, file_id=res_file.pk,
                                               fail_feedback=False)
                     elif res_file.has_logical_file and file_type is not None and \
                             not isinstance(res_file.logical_file, file_type):
@@ -170,7 +526,8 @@ def __ingest_irods_directory(self,
                             raise ValidationError(msg)
 
         for dname in listing[0]:  # directories
-            error2, ecount2 = __ingest_irods_directory(self,
+            # do not use os.path.join because fname might contain unicode characters
+            error2, ecount2 = __ingest_irods_directory(resource,
                                                        dir + '/' + dname,
                                                        logger,
                                                        stop_on_error=stop_on_error,
@@ -188,7 +545,12 @@ def __ingest_irods_directory(self,
 
 
 def check_for_dangling_irods(echo_errors=True, log_errors=False, return_errors=False):
-    """ This checks for resource trees in iRODS with no correspondence to Django at all """
+    """ This checks for resource trees in iRODS with no correspondence to Django at all
+
+    :param log_errors: whether to log errors to Django log
+    :param echo_errors: whether to print errors on stdout
+    :param return_errors: whether to collect errors in an array and return them.
+    """
 
     istorage = IrodsStorage()  # local only
     toplevel = istorage.listdir('.')  # list the resources themselves
@@ -235,7 +597,60 @@ class CheckJSONLD(object):
             return
 
 
+def repair_resource(resource, logger, stop_on_error=False,
+                    log_errors=True, echo_errors=False, return_errors=False):
+    errors = []
+    ecount = 0
+
+    try:
+        resource = get_resource_by_shortkey(resource.short_id, or_404=False)
+    except BaseResource.DoesNotExist:
+        msg = "Resource with id {} not found in Django Resources".format(resource.short_id)
+        if log_errors:
+            logger.error(msg)
+        if echo_errors:
+            print(msg)
+        if return_errors:
+            errors.append(msg)
+            ecount = ecount + 1
+        return errors, ecount
+
+    print("REPAIRING RESOURCE {}".format(resource.short_id))
+
+    # ingest any dangling iRODS files that you can
+    # Do this before check because otherwise, errors get printed twice
+    # TODO: This does not currently work properly for composite resources
+    # if resource.resource_type == 'CompositeResource' or \
+    if resource.resource_type == 'GenericResource' or \
+       resource.resource_type == 'ModelInstanceResource' or \
+       resource.resource_type == 'ModelProgramResource':
+        _, count = ingest_irods_files(resource,
+                                      logger,
+                                      stop_on_error=False,
+                                      echo_errors=True,
+                                      log_errors=False,
+                                      return_errors=False)
+        if count:
+            print("... affected resource {} has type {}, title '{}'"
+                  .format(resource.short_id, resource.resource_type,
+                          resource.title.encode('string-escape')))
+
+    _, count = check_irods_files(resource,
+                                 stop_on_error=False,
+                                 echo_errors=True,
+                                 log_errors=False,
+                                 return_errors=False,
+                                 clean_irods=False,
+                                 clean_django=True,
+                                 sync_ispublic=True)
+    if count:
+        print("... affected resource {} has type {}, title '{}'"
+              .format(resource.short_id, resource.resource_type,
+                      resource.title.encode('string-escape')))
+
+
 class CheckResource(object):
+    """ comprehensively check a resource """
     header = False
 
     def __init__(self, short_id):
@@ -264,13 +679,10 @@ class CheckResource(object):
         # print("TESTING {}".format(self.short_id))  # leave this for debugging
 
         try:
-            res = BaseResource.objects.get(short_id=self.short_id)
+            self.resource = get_resource_by_shortkey(self.short_id, or_404=False)
         except BaseResource.DoesNotExist:
             print("{} does not exist in Django".format(self.short_id))
             return
-
-        self.resource = res.get_content_model()
-        assert self.resource, (res, res.content_model)
 
         # skip federated resources if not configured to handle these
         if self.resource.is_federated and not settings.REMOTE_USE_IRODS:
@@ -283,23 +695,27 @@ class CheckResource(object):
         if not istorage.exists(self.resource.root_path):
             self.label()
             print("  root path {} does not exist in iRODS".format(self.resource.root_path))
-            print("  ... resource {} has type {}".format(self.resource.short_id,
-                                                         self.resource.resource_type))
+            print("  ... resource {} has type {} and title {}"
+                  .format(self.resource.short_id,
+                          self.resource.resource_type,
+                          self.resource.title.encode('string-escape')))
             return
 
         for a in ('bag_modified', 'isPublic', 'resourceType', 'quotaUserName'):
             value = self.check_avu(a)
             if a == 'resourceType' and value is not None and value != self.resource.resource_type:
                 self.label()
-                print("  AVU resourceType is {}, should be {}".format(value,
+                print("  AVU resourceType is {}, should be {}".format(value.encode('string-escape'),
                                                                       self.resource.resource_type))
             if a == 'isPublic' and value is not None and value != self.resource.raccess.public:
                 self.label()
-                print("  AVU isPublic is {}, but public is {}".format(value,
+                print("  AVU isPublic is {}, but public is {}".format(value.encode('string-escape'),
                                                                       self.resource.raccess.public))
 
-        irods_issues, irods_errors = self.resource.check_irods_files(log_errors=False,
-                                                                     return_errors=True)
+        irods_issues, irods_errors = check_irods_files(self.resource,
+                                                       log_errors=False,
+                                                       echo_errors=False,
+                                                       return_errors=True)
 
         if irods_errors:
             self.label()
